requirements.txt
flask
uvicorn
python-dotenv
fastapi[standard]
redis
pika
stomp.py